+++ category_encoders/leave_one_out.py
@@ -260,7 +260,7 @@ class LeaveOneOutEncoder(BaseEstimator, TransformerMixin):
             level_notunique = colmap['count'] > 1
 
             unique_train = colmap.index
-            unseen_values = pd.Series([x for x in X[col].unique() if x not in unique_train])
+            unseen_values = pd.Series([x for x in X[col].unique() if x not in unique_train], dtype=unique_train.dtype)
 
             is_nan = X[col].isnull()
             is_unknown_value = X[col].isin(unseen_values.dropna().astype(object))
